requirements.txt
opencv-python
numpy
pypng
Pillow
fastapi[all]
ultralytics
torch